% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{divisions}
\alias{divisions}
\title{NCAAF team names and the conferences and divisions they belong to}
\format{
A data frame with 130 rows and 4 variables containing NCAAF team level
information:
\describe{
\item{team}{Team name}
\item{conf}{Conference abbreviation}
\item{division}{Division name}
\item{sdiv}{Division abbreviation}
}
This data frame is created using the \code{cfbd_team_info} function and the
\code{cfbd_conf_types_df} data frame of the \code{cfbfastR} package.
Please see \code{data-raw/divisions.R} for the code to create this data.
}
\usage{
divisions
}
\description{
NCAAF team names and the conferences and divisions they belong to
}
\examples{
\donttest{
divisions
}
}
\keyword{datasets}
